remove "datasets/bySubject/toMonth" #4
We are removing the "datasets/bySubject/toMonth" endpoint from
the sample config and README because it currently requires that
all Dataverse installations be 4.11 or higher. We want to keep
the bar lower and focus for now on showing the 6 plots that only
require Dataverse 4.9 or higher.

We'll re-open #3 to track this 7th endpoint.